+++ src/objects.js
@@ -94,7 +94,7 @@ embedMetadataPNG*/
 
 /*jshint esversion: 6*/
 
-modules.objects = '2022-April-28';
+modules.objects = '2022-May-02';
 
 var SpriteMorph;
 var StageMorph;
@@ -1103,7 +1103,12 @@ SpriteMorph.prototype.initBlocks = function () {
             type: 'command',
             category: 'sensing',
             spec: 'set %byob of block %repRing to %s',
-            defaults: [['definition']]
+            defaults: [['label']]
+        },
+        reportDefineBlock: {
+            type: 'reporter',
+            category: 'sensing',
+            spec: 'define block %s %repRing'
         },
 
         // Operators
@@ -2678,6 +2683,7 @@ SpriteMorph.prototype.blockTemplates = function (
         blocks.push(block('doResetTimer'));
         blocks.push(watcherToggle('getTimer'));
         blocks.push(block('getTimer'));
+        blocks.push(block('reportDate'));
         blocks.push('-');
         blocks.push(block('reportAttributeOf'));
 
@@ -2695,9 +2701,9 @@ SpriteMorph.prototype.blockTemplates = function (
         blocks.push(block('reportGlobalFlag'));
         blocks.push(block('doSetGlobalFlag'));
         blocks.push('-');
-        blocks.push(block('reportDate'));
         blocks.push(block('reportBlockAttribute'));
         blocks.push(block('doSetBlockAttribute'));
+        blocks.push(block('reportDefineBlock'));
 
         // for debugging: ///////////////
         if (devMode) {
@@ -9146,6 +9152,7 @@ StageMorph.prototype.blockTemplates = function (
         blocks.push(block('doResetTimer'));
         blocks.push(watcherToggle('getTimer'));
         blocks.push(block('getTimer'));
+        blocks.push(block('reportDate'));
         blocks.push('-');
         blocks.push(block('reportAttributeOf'));
 
@@ -9163,9 +9170,9 @@ StageMorph.prototype.blockTemplates = function (
         blocks.push(block('reportGlobalFlag'));
         blocks.push(block('doSetGlobalFlag'));
         blocks.push('-');
-        blocks.push(block('reportDate'));
         blocks.push(block('reportBlockAttribute'));
         blocks.push(block('doSetBlockAttribute'));
+        blocks.push(block('reportDefineBlock'));
 
         // for debugging: ///////////////
         if (this.world().isDevMode) {